+++ builders/deploy.yaml
@@ -33,12 +33,12 @@ steps:
   args:
   - publish
   - -P
-  - ./cmd/wipeout-export
+  - ./cmd/cleanup-export
 - name: 'gcr.io/$PROJECT_ID/ko'
   args:
   - publish
   - -P
-  - ./cmd/wipeout-exposure
+  - ./cmd/cleanup-exposure
 # Deploy to Cloud Run.
 - name: 'gcr.io/cloud-builders/gcloud'
   entrypoint: 'bash'
